mg.1: Add delete/kill/yank/undo subsection
Introduce "kill zone" term.  "Kill buffer" would be misleading because
it's not a buffer in the Emacs sense; you can't open a window on it (and
where would text killed from the kill buffer go?).  Also, there is a
`kill-buffer` command that...kills a buffer.

"Kill ring" is also plainly incorrect since it's not a ring, but a
single datum.