extend.c (load): Make startup errors visible
Fix a problem where, for instance, invalid ~/.mg contents would cause a
beep/flash at startup but no error message was visible.  It was also not
made clear whether mg kept trying to read the file, or abandoned it upon
error.  Make it clearer what is going on (the latter), and halt the user
in the minibuffer when this happens.