paragraph.c: Make set-fill-column more useful
Model it on the new `set-prompt-delay`.

Report the current fill column when prompting to set a new one.  (GNU
Emacs does that too.)  Since there is no other way to query it, this
helps the user to decide to C-g out if they're happy with the value.
Accept negative inputs in the minibuffer.  Diagnose nonpositive values;
this way we issue a message whether the out-of-range value came from the
minibuffer prompt or a command parameter.  Parallelize wording upon
success with `set-prompt-delay`, `audible-bell`, and `visible-bell`.