Drop return type from `setlineno()`
The function cannot fail, and returns no information.  It seems to bear
a return type only to support terser source code like
"return(setlineno(n));", but I find that to be a dubious optimization.
It's more important in my opinion to indicate where data is being
passed.  And if one wants to quibble, I reckon this approach clobbers
one fewer register (or requires one fewer word on the stack).

* def.h: Update declaration.
* basic.c (setlineno): Update definition.
* basic.c (gotoline):
* buffer.c (dorevert): Update call sites.